// app/interfaces/dto/user-filter.dto.ts
export interface UserFilterDto {
    filterValue: string;
    filterType: string;
    ageRange: string;
    userActive: boolean;
    birthDateBegin: string;
    birthDateEnd: string;
    registerDateBegin: string;
    registerDateEnd: string;
    updateDateBegin: string;
    updateDateEnd: string;
}